+++ sliceutil/sliceutil.go
@@ -1,4 +1,9 @@
-// Package sliceutil provides a set of functions for working with slices.
+// Package sliceutil provides functions for working with slices.
+//
+// The "set" helpers are simple implementations, and don't operate on true
+// "sets" (e.g. it will retain order, []int64 can contain duplicates). Consider
+// using something like golang-set if you want to use sets and care a lot about
+// performance: https://github.com/deckarep/golang-set
 package sliceutil // import "github.com/teamwork/utils/sliceutil"
 
 import (
@@ -120,28 +125,6 @@ func RepeatString(s string, n int) (r []string) {
 		r = append(r, s)
 	}
 	return r
-}
-
-// Difference returns a new slice with elements that are in "set" but not in
-// "others".
-func Difference(set []int64, others ...[]int64) []int64 {
-	out := []int64{}
-
-	for _, setItem := range set {
-		found := false
-		for _, o := range others {
-			if InInt64Slice(o, setItem) {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			out = append(out, setItem)
-		}
-	}
-
-	return out
 }
 
 // Range creates an []int counting at "start" up to (and including) "end".
